Register events with a specific priority. Close #1339
