+++ src/main/java/com/blamejared/crafttweaker/impl/events/CTEventManager.java
@@ -1,10 +1,9 @@
 package com.blamejared.crafttweaker.impl.events;
 
 import com.blamejared.crafttweaker.api.CraftTweakerAPI;
-import com.blamejared.crafttweaker.api.actions.IUndoableAction;
 import com.blamejared.crafttweaker.api.annotations.ZenRegister;
+import com.blamejared.crafttweaker.impl.actions.events.ActionRegisterEvent;
 import com.blamejared.crafttweaker_annotations.annotations.Document;
-import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.Event;
 import net.minecraftforge.eventbus.api.EventPriority;
 import org.openzen.zencode.java.ZenCodeType;
@@ -42,32 +41,27 @@ public class CTEventManager {
      */
     @ZenCodeType.Method
     public static <T extends Event> void register(Class<T> typeOfT, Consumer<T> consumer) {
-        CraftTweakerAPI.apply(new IUndoableAction() {
-            private final EventHandlerWrapper<T> eventHandler = new EventHandlerWrapper<T>(consumer);
-    
-            @Override
-            public void undo() {
-        
-                MinecraftForge.EVENT_BUS.unregister(eventHandler);
-            }
-    
-            @Override
-            public String describeUndo() {
-        
-                return "Unregistering event listener for " + typeOfT.getSimpleName() + ".";
-            }
-            
-            @Override
-            public void apply() {
-                //Let's go completely safe and use the type
-                MinecraftForge.EVENT_BUS.addListener(EventPriority.NORMAL, false, typeOfT, eventHandler);
-            }
-            
-            @Override
-            public String describe() {
-                return "Registering event listener for " + typeOfT.getSimpleName() + ".";
-            }
-        });
+        register(typeOfT, EventPriority.NORMAL, consumer);
     }
-    
+
+
+    /**
+     * Registers a new Event listener with a specific priority.
+     * @param typeOfT Internally used to determine the Event, invisible to scripts.
+     * @param priority priority for this listener
+     * @param consumer The event handler as consumer
+     * @param <T> The type of the event
+     *
+     * @docParam <T> crafttweaker.api.event.entity.player.MCAnvilRepairEvent
+     * @docParam priority EventPriority.HIGHEST
+     * @docParam consumer (event) => {
+     *     var player = event.player;
+     *     var result = event.itemResult;
+     *     println("Player '" + player.name + "' crafted " + result.commandString);
+     * }
+     */
+    @ZenCodeType.Method
+    public static <T extends Event> void register(Class<T> typeOfT, EventPriority priority, Consumer<T> consumer) {
+        CraftTweakerAPI.apply(new ActionRegisterEvent<>(typeOfT, consumer, priority));
+    }
 }
